[feat]:Medicamentos Actuales listo

+++ src/presentation/services/Paciente/MedicacionActualService.ts
@@ -1,3 +1,5 @@
+import { Admision } from "../../../data/models/Admision";
+import { Medicamentos } from "../../../data/models/Medicamentos";
 import { Paciente_Medicacion_Actual } from "../../../data/models/Paciente_Medicacion_Actual";
 import { createMedicacionActualDto } from "../../../domain/Dtos/pacientes/Medicacion Actual/createMedicacionActualDto";
 import { updateMedicacionActualDto } from "../../../domain/Dtos/pacientes/Medicacion Actual/updateMedicacionActualDto";
@@ -7,14 +9,28 @@ import { MedicamentosServices } from "../MedicamentosServices";
 
 export class MedicacionActualService {
 
-    public static async buscarLasMedicacionesActualesPorPaciente(id_Paciente: number):Promise<[string?,Paciente_Medicacion_Actual[]?]>{//todo:testear
+    public static async buscarLasMedicacionesActualesPorPacienteYAdmision(id_Paciente: number, id_Admision: number):Promise<[string?,Paciente_Medicacion_Actual[]?]>{//todo:testear
         try {
             
             if(!id_Paciente || Number(id_Paciente) < 0) return ["id_Paciente es nulo o menor que 0",undefined]
+            if(!id_Admision || Number(id_Admision) < 0) return ["id_Admision es nulo o menor que 0",undefined]
             const medicacionesPaciente = await Paciente_Medicacion_Actual.findAll({
                 where: {
-                    id_Paciente: id_Paciente
-                }
+                    id_Paciente: id_Paciente,
+                    id_Admision: id_Admision
+                },
+                include:[
+                    {
+                        model: Admision,
+                        as: "admision",
+                        attributes: ["fecha_de_admision"]
+                    },
+                    {
+                        model: Medicamentos,
+                        as: "medicamentos",
+                        attributes: ["nombre"]    
+                    }
+                ]
             })
             if(!medicacionesPaciente) return ["No hay registradas medicaciones actuales",undefined]
             return [undefined, medicacionesPaciente]
@@ -40,10 +56,16 @@ export class MedicacionActualService {
             return [error as string,undefined]
         }
     }
-    public static async existeMedicacionActual(_createMediacionActual: createMedicacionActualDto):Promise<[string?,boolean?]>{//todo:testear
+    public static async existeMedicacionActual(_createMedicacionActual?: createMedicacionActualDto, _updateMedicacionActual?:updateMedicacionActualDto):Promise<[string?,boolean?]>{//todo:testear
         try {
+            if(!_createMedicacionActual && !_updateMedicacionActual) return ["Se requiere el _createMedicacionActual || _updateMedicacionActual "]
+            const objectMedicacionActual = (_createMedicacionActual)? createMedicacionActualDto.toObject(_createMedicacionActual): updateMedicacionActualDto.toObject(_updateMedicacionActual)
             const confirmacion = await Paciente_Medicacion_Actual.findOne({
-                where: createMedicacionActualDto.toObject(_createMediacionActual)
+                where: {
+                    id_Medicamento : objectMedicacionActual.id_Medicamento,
+                    id_Paciente : objectMedicacionActual.id_Paciente,
+                    id_Admision : objectMedicacionActual.id_Admision
+                }
             })
             if(confirmacion) return [undefined,true]
             return [undefined, false]
@@ -56,7 +78,7 @@ export class MedicacionActualService {
         try {
             
             if(!await MedicamentosServices.buscarMedicamentoPorId(_createMedicacionActual.id_Medicamento).then(res => res[1])) return["No se encontro el medicamento registrado por dicho Id", undefined]
-            if(await this.existeMedicacionActual(_createMedicacionActual).then(res=> res[1])){
+            if(await this.existeMedicacionActual(_createMedicacionActual, null  ).then(res=> res[1])){
                 return ["Ya existe dicho registro", undefined]
             }
             const medicamentoCreado = await Paciente_Medicacion_Actual.create(createMedicacionActualDto.toObject(_createMedicacionActual))
@@ -69,7 +91,11 @@ export class MedicacionActualService {
     }
     public static async actualizarMedicacionActual(_updateMedicacionActual: updateMedicacionActualDto):Promise<[string?,boolean?]>{//todo:testear
         try {
-            if(!await this.buscarMedicacionActualPorId(_updateMedicacionActual.id_Paciente_Medicacion_Actual).then(res=> res[1])) return ["No se encontro registrada la mediacion actual por id", false]
+            const MedicacionActual = await this.buscarMedicacionActualPorId(_updateMedicacionActual.id_Paciente_Medicacion_Actual)
+            if(!MedicacionActual[1]) return ["No se encontro registrada la medicacion actual por id", false]
+            MedicacionActual[1].id_Medicamento = _updateMedicacionActual.id_Medicamento
+            
+            if(await this.existeMedicacionActual(null,_updateMedicacionActual).then(res=>res[1])) return ["Ya existe dicho registro", false]
             if(!await MedicamentosServices.buscarMedicamentoPorId(_updateMedicacionActual.id_Medicamento).then(res=>res[1])) return ["No se encontro el medicamento registrado con dicho Id"]
             const actualizacion = await Paciente_Medicacion_Actual.update(updateMedicacionActualDto.toObject(_updateMedicacionActual), {
                 where: {
@@ -86,6 +112,7 @@ export class MedicacionActualService {
     }
     public static async eliminarMedicacionActual(id_Paciente_Medicacion_Actual:number):Promise<[string?,boolean?]>{//todo:testear
         try {
+            if(!id_Paciente_Medicacion_Actual || Number(id_Paciente_Medicacion_Actual) < 0) return ["id_Paciente_Medicacion_Actual es nulo o menor que 0", false]
             if(!this.buscarMedicacionActualPorId(id_Paciente_Medicacion_Actual).then(res=> res[1])) return ["No se encontro la medicacion actual por id registrada", false]
             const confirmacion = await Paciente_Medicacion_Actual.destroy({
                 where: {
